Change the download mock to have an async completion
