+++ DuckDuckGoTests/DownloadsManagerTests.swift
@@ -21,55 +21,7 @@ import XCTest
 import WebKit
 
 @testable import DuckDuckGo
-
-private struct MockSessionSetup {
-    let tmpFinalPath: URL
-    let documentsFinalPath: URL
-    let session: MockDownloadSession
-    let response: MockNavigationResponse
-    
-    init(mimeType: String, downloadsManager: DownloadsManager) {
-        let tmpName = "MOCK_\(UUID().uuidString).tmp"
-        let filename = "\(UUID().uuidString).zip"
-        
-        response = MockNavigationResponse()
-        response.suggestedFileName = filename
-        response.mimeType = mimeType
-        
-        session = MockDownloadSession(DownloadTestsHelper.mockURL)
-
-        let tmpPath = DownloadTestsHelper.tmpDirectory.appendingPathComponent(tmpName)
-        
-        tmpFinalPath = DownloadTestsHelper.tmpDirectory.appendingPathComponent(filename)
-        documentsFinalPath = DownloadTestsHelper.documentsDirectory.appendingPathComponent(filename)
-        
-        session.temporaryFilePath = tmpPath
-        
-        DownloadTestsHelper.createMockFile(on: tmpPath)
-    }
-}
-
-class MockURLResponse: URLResponse {
-    var mockFileName: String?
-    
-    override var suggestedFilename: String? {
-        mockFileName
-    }
-}
-
-class MockNavigationResponse: WKNavigationResponse {
-    var suggestedFileName: String?
-    var mimeType: String?
-    
-    override var response: URLResponse {
-        let response = MockURLResponse(url: URL(string: "https://www.duck.com")!,
-                        mimeType: mimeType!,
-                        expectedContentLength: 1234,
-                        textEncodingName: "")
-        response.mockFileName = suggestedFileName
-        return response
-    }
-}
+import WidgetKit
 
 class DownloadsManagerTests: XCTestCase {
     var mockDependencyProvider: MockDependencyProvider!
@@ -82,61 +34,109 @@ class DownloadsManagerTests: XCTestCase {
     override func tearDown() {
         AppDependencyProvider.shared = AppDependencyProvider()
         DownloadTestsHelper.deleteAllFiles()
+        // swiftlint:disable notification_center_detachment
+        NotificationCenter.default.removeObserver(self)
+        // swiftlint:enable notification_center_detachment
     }
-
+    
     func testTemporaryPKPassDownload() {
         let downloadsManager = mockDependencyProvider.downloadsManager
-
+        
         let sessionSetup = MockSessionSetup(mimeType: "application/vnd.apple.pkpass", downloadsManager: downloadsManager)
-    
+        
         let download = downloadsManager.setupDownload(sessionSetup.response, downloadSession: sessionSetup.session)!
         XCTAssertTrue(download.temporary, "Download should be temporary")
         
-        downloadsManager.startDownload(download)
+        let expectation = expectation(description: "Download finish")
         
-        XCTAssertTrue(DownloadTestsHelper.checkIfFileExists(sessionSetup.tmpFinalPath), "File should exist")
-        XCTAssertFalse(DownloadTestsHelper.checkIfFileExists(sessionSetup.documentsFinalPath), "File should not exist")
+        NotificationCenter.default.addObserver(forName: .downloadFinished, object: nil, queue: nil) { notification in
+            if DownloadTestsHelper.downloadForNotification(notification) == download {
+                XCTAssertTrue(DownloadTestsHelper.checkIfFileExists(sessionSetup.tmpFinalPath), "File should exist")
+                XCTAssertFalse(DownloadTestsHelper.checkIfFileExists(sessionSetup.documentsFinalPath), "File should not exist")
+                expectation.fulfill()
+            }
+        }
+        
+        downloadsManager.startDownload(download)
+        wait(for: [expectation], timeout: 1)
     }
     
     func testTemporaryRealityDownload() {
         let downloadsManager = mockDependencyProvider.downloadsManager
-
+        
         let sessionSetup = MockSessionSetup(mimeType: "model/vnd.reality", downloadsManager: downloadsManager)
-    
+        
         let download = downloadsManager.setupDownload(sessionSetup.response, downloadSession: sessionSetup.session)!
         XCTAssertTrue(download.temporary, "Download should be temporary")
         
-        downloadsManager.startDownload(download)
+        let expectation = expectation(description: "Download finish")
         
-        XCTAssertTrue(DownloadTestsHelper.checkIfFileExists(sessionSetup.tmpFinalPath), "File should exist")
-        XCTAssertFalse(DownloadTestsHelper.checkIfFileExists(sessionSetup.documentsFinalPath), "File should not exist")
+        NotificationCenter.default.addObserver(forName: .downloadFinished, object: nil, queue: nil) { notification in
+            if DownloadTestsHelper.downloadForNotification(notification) == download {
+                XCTAssertTrue(DownloadTestsHelper.checkIfFileExists(sessionSetup.tmpFinalPath), "File should exist")
+                XCTAssertFalse(DownloadTestsHelper.checkIfFileExists(sessionSetup.documentsFinalPath), "File should not exist")
+                expectation.fulfill()
+            }
+        }
+        
+        downloadsManager.startDownload(download)
+        wait(for: [expectation], timeout: 1)
+        
     }
     
     func testTemporaryUSDZDownload() {
         let downloadsManager = mockDependencyProvider.downloadsManager
-
+        
         let sessionSetup = MockSessionSetup(mimeType: "model/vnd.usdz+zip", downloadsManager: downloadsManager)
-    
+        
         let download = downloadsManager.setupDownload(sessionSetup.response, downloadSession: sessionSetup.session)!
         XCTAssertTrue(download.temporary, "Download should be temporary")
         
-        downloadsManager.startDownload(download)
+        let expectation = expectation(description: "Download finish")
         
-        XCTAssertTrue(DownloadTestsHelper.checkIfFileExists(sessionSetup.tmpFinalPath), "File should exist")
-        XCTAssertFalse(DownloadTestsHelper.checkIfFileExists(sessionSetup.documentsFinalPath), "File should not exist")
+        NotificationCenter.default.addObserver(forName: .downloadFinished, object: nil, queue: nil) { notification in
+            if DownloadTestsHelper.downloadForNotification(notification) == download {
+                XCTAssertTrue(DownloadTestsHelper.checkIfFileExists(sessionSetup.tmpFinalPath), "File should exist")
+                XCTAssertFalse(DownloadTestsHelper.checkIfFileExists(sessionSetup.documentsFinalPath), "File should not exist")
+                expectation.fulfill()
+            }
+        }
+        
+        downloadsManager.startDownload(download)
+        wait(for: [expectation], timeout: 1)
     }
-
+    
     func testPermanentBinaryDownload() {
         let downloadsManager = mockDependencyProvider.downloadsManager
-
+        
         let sessionSetup = MockSessionSetup(mimeType: "application/octet-stream", downloadsManager: downloadsManager)
-    
+        
         let download = downloadsManager.setupDownload(sessionSetup.response, downloadSession: sessionSetup.session)!
         XCTAssertFalse(download.temporary, "download should not be temporary")
         
-        downloadsManager.startDownload(download)
+        let expectation = expectation(description: "Download finish")
         
-        XCTAssertFalse(DownloadTestsHelper.checkIfFileExists(sessionSetup.tmpFinalPath), "File should not exist")
-        XCTAssertTrue(DownloadTestsHelper.checkIfFileExists(sessionSetup.documentsFinalPath), "File should exist")
+        NotificationCenter.default.addObserver(forName: .downloadFinished, object: nil, queue: nil) { notification in
+            if DownloadTestsHelper.downloadForNotification(notification) == download {
+                
+                XCTAssertFalse(DownloadTestsHelper.checkIfFileExists(sessionSetup.tmpFinalPath), "File should not exist")
+                XCTAssertTrue(DownloadTestsHelper.checkIfFileExists(sessionSetup.documentsFinalPath), "File should exist")
+                expectation.fulfill()
+            }
+        }
+        
+        downloadsManager.startDownload(download)
+        wait(for: [expectation], timeout: 1)
+    }
+    
+    func testIfFinishedDownloadIsRemovedFromList() {
+        
+    }
+    
+    func downloadForNotification(_ notification: Notification) -> Download {
+        if let download = notification.userInfo?[DownloadsManager.UserInfoKeys.download] as? Download {
+            return download
+        }
+        fatalError("Should only be used to test valid downloads")
     }
 }
