--- DuckDuckGoTests/MockDownloadSession.swift
@@ -1,35 +1,0 @@
-//
-//  MockDownloadSession.swift
-//  DuckDuckGo
-//
-//  Copyright © 2022 DuckDuckGo. All rights reserved.
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//  http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-//
-
-import Foundation
-
-@testable import DuckDuckGo
-
-class MockDownloadSession: DownloadSession {
-    var temporaryFilePath: URL?
-    var error: Error?
-    
-    override func start() {
-        let session = URLSession.shared
-        let task = session.downloadTask(with: URL(string: "https://duck.com")!)
-
-        delegate?.urlSession(session, downloadTask: task, didFinishDownloadingTo: temporaryFilePath!)
-        delegate?.urlSession(URLSession.shared, task: task, didCompleteWithError: error)
-    }
-}
